Make JObject::oid_t unsigned long long as per Paul M.'s suggestion
git-svn-id: https://phys12svn.jlab.org/repos/JANA@2434 c5ed4466-e916-0410-8347-b3263e9c103d

+++ src/JANA/JObject.h
@@ -60,7 +60,7 @@ class JObject{
 	public:
 		JOBJECT_PUBLIC(JObject);
 
-		typedef unsigned long oid_t;
+		typedef unsigned long long oid_t;
 	
 		JObject() : id((oid_t)this),append_types(false),factory(NULL) {}
 		JObject( oid_t aId ) : id( aId ),append_types(false),factory(NULL) {}
